Strip ANSI escape codes from log lines before validating and printing
Trying to run the regex against a log line string which has escape codes
could lead to unexpected behavior from the user and more difficult to
write validation rules.

Additionally, if we're going to include our own escape codes for
coloring in the future, colorings in the log line itself could be
distracting.